% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{ColoresRacafe}
\alias{ColoresRacafe}
\title{Generar paleta de colores corporativos Racafe}
\usage{
ColoresRacafe(...)
}
\arguments{
  \item{n}{Numero de colores requeridos. Maximo 10.}
}
\value{Vector de colores hexadecimales.}
\description{
Generar paleta de colores corporativos Racafe
}
\examples{
ColoresRacafe(5)
ColoresRacafe(1)
}
